backend/requirements.txt
pdfplumber
pdf2image
Pillow
opencv-python
fastapi
uvicorn
python-multipart
